refactor: 컨텍스트 메뉴 위치 함수 utils로 빼기
